+++ citation_pulse/__init__.py
@@ -2,4 +2,4 @@
 CitationPulse package initialization.
 """
 
-__version__ = "1.0.0"
+__version__ = "1.2.0"